Test fixes, fix writeTxt; delete bad test, move some others
