+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/linalg/Nd4jTestsC.java
@@ -165,16 +165,6 @@ public class Nd4jTestsC extends BaseNd4jTest {
     public void testDiag() {
       INDArray diag = Nd4j.diag(Nd4j.linspace(1,4,4).reshape(4,1));
       assertArrayEquals(new long[] {4,4},diag.shape());
-
-    }
-
-    @Test
-    public void testSoftmaxDerivativeGradient() {
-        INDArray input = Nd4j.linspace(1,4,4).reshape(2,2);
-        INDArray inputDup = input.dup();
-        Nd4j.getExecutioner().exec(new org.nd4j.linalg.api.ops.impl.transforms.gradient.SoftMaxDerivative(input,Nd4j.ones(2,2),input));
-        Nd4j.getExecutioner().exec(new SoftMaxDerivative(inputDup));
-        assertEquals(input,inputDup);
     }
 
     @Test
@@ -6488,48 +6478,6 @@ public class Nd4jTestsC extends BaseNd4jTest {
         log.info("Result shape: {}", result.shapeInfoDataBuffer().asLong());
 
         Nd4j.setDataType(dtype);
-    }
-
-
-    @Test
-    public void testEye(){
-
-        int[] rows = new int[]{3,3,3,3};
-        int[] cols = new int[]{3,2,2,2};
-        int[][] batch = new int[][]{null, null, {4}, {3,3}};
-        INDArray[] expOut = new INDArray[4];
-
-        expOut[0] = Nd4j.eye(3);
-        expOut[1] = Nd4j.create(new double[][]{{1,0},{0,1}, {0, 0}});
-        expOut[2] = Nd4j.create(4,3,2);
-        for( int i=0; i<4; i++ ){
-            expOut[2].get(NDArrayIndex.point(i), NDArrayIndex.all(), NDArrayIndex.all()).assign(expOut[1]);
-        }
-        expOut[3] = Nd4j.create(3,3,3,2);
-        for( int i=0; i<3; i++ ){
-            for( int j=0; j<3; j++ ) {
-                expOut[3].get(NDArrayIndex.point(i), NDArrayIndex.point(j), NDArrayIndex.all(), NDArrayIndex.all()).assign(expOut[1]);
-            }
-        }
-
-
-        for(int i=0; i<3; i++ ) {
-            INDArray out = Nd4j.create(expOut[i].shape());
-
-            val opB = DynamicCustomOp.builder("eye")
-                    .addOutputs(out)
-                    .addIntegerArguments(rows[i], cols[i]);
-
-            if(batch[i] != null)
-                opB.addIntegerArguments(batch[i]);
-
-
-            val op = opB.build();
-
-            Nd4j.getExecutioner().exec(op);
-
-            assertEquals(expOut[i], out);
-        }
     }
 
     @Test
