+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/linalg/api/ndarray/TestNdArrReadWriteTxt.java
@@ -2,7 +2,9 @@ package org.nd4j.linalg.api.ndarray;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.ArrayUtils;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.nd4j.linalg.BaseNd4jTest;
@@ -11,6 +13,7 @@ import org.nd4j.linalg.factory.Nd4j;
 import org.nd4j.linalg.factory.Nd4jBackend;
 import org.nd4j.linalg.primitives.Pair;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
@@ -26,23 +29,28 @@ import static org.junit.Assert.assertEquals;
 @RunWith(Parameterized.class)
 public class TestNdArrReadWriteTxt extends BaseNd4jTest {
 
+
+    @Rule
+    public TemporaryFolder testDir = new TemporaryFolder();
+
     public TestNdArrReadWriteTxt(Nd4jBackend backend) {
         super(backend);
     }
 
     @Test
-    public void compareAfterWrite() {
+    public void compareAfterWrite() throws Exception {
         int [] ranksToCheck = new int[] {0,1,2,3,4};
         for (int i=0; i<ranksToCheck.length;i++) {
             log.info("Checking read write arrays with rank " + ranksToCheck[i]);
-            compareArrays(ranksToCheck[i],ordering());
+            compareArrays(ranksToCheck[i],ordering(), testDir);
         }
     }
 
-    public static void compareArrays(int rank, char ordering) {
+    public static void compareArrays(int rank, char ordering, TemporaryFolder testDir) throws Exception {
         List<Pair<INDArray, String>> all = NDArrayCreationUtil.getTestMatricesWithVaryingShapes(rank,ordering);
         Iterator<Pair<INDArray,String>> iter = all.iterator();
         while (iter.hasNext()) {
+            File dir = testDir.newFolder();
             Pair<INDArray,String> currentPair = iter.next();
             INDArray origArray = currentPair.getFirst();
             //adding elements outside the bounds where print switches to scientific notation
@@ -50,14 +58,9 @@ public class TestNdArrReadWriteTxt extends BaseNd4jTest {
             origArray.putScalar(0,10001.1234);
             log.info("\nChecking shape ..." + currentPair.getSecond());
             log.info("\n"+ origArray.dup('c').toString());
-            Nd4j.writeTxt(origArray, "someArr.txt");
-            INDArray readBack = Nd4j.readTxt("someArr.txt");
+            Nd4j.writeTxt(origArray, new File(dir, "someArr.txt").getAbsolutePath());
+            INDArray readBack = Nd4j.readTxt(new File(dir, "someArr.txt").getAbsolutePath());
             assertEquals("\nNot equal on shape " + ArrayUtils.toString(origArray.shape()), origArray, readBack);
-            try {
-                Files.delete(Paths.get("someArr.txt"));
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
         }
     }
